style : style

+++ main-service/src/main/java/ru/practicum/ewm/dto/event/NewEventDto.java
@@ -31,7 +31,7 @@ public class NewEventDto {
     @NotNull(message = MESSAGE_VALIDATION_NOT_BLANK)
     private Location location;
     private Boolean paid = false;
-    private Integer participantLimit = 0;
+    private Long participantLimit = 0L;
     private Boolean requestModeration = true;
     @NotBlank(message = MESSAGE_VALIDATION_NOT_BLANK)
     @Size(min = 3, max = 120, message = MESSAGE_VALIDATION_SIZE)
